Remove crates/query-proxy
We no longer have the flags proxy so there's no sense to share
definitions.
