--- crates/query-proxy/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "query-proxy"
-version = "0.1.0"
-authors = ["Blaž Hrastnik <blaz@smartcontract.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[features]
-# for more explicit tests, cargo test --features=backtraces
-backtraces = ["cosmwasm-std/backtraces"]
-library = []
-
-[dependencies]
-cosmwasm-std = { version = "0.16.2" }
-cosmwasm-storage = { version = "0.16.0" }
-cw-storage-plus = "0.9.0"
-cw2 = "0.9.0"
-schemars = "0.8.3"
-serde = { version = "1.0.103", default-features = false, features = ["derive"] }
-thiserror = { version = "1.0.24" }
-owned = { path = "../../crates/owned", default-features = false, features = ["library"] }
-
-[dev-dependencies]
-cosmwasm-schema = { version = "0.16.0" }
-cw-multi-test = "0.9.1"
